fix: adding a parent link and a user member description
BREAKING CHANGE: adding a parent link and a user member description

fix #43
