+++ src/pages/page.jsx
@@ -1,0 +1,12 @@
+import ArticleCard from "@/components/ArticleCard";
+import React from "react";
+
+const page = () => {
+    return (
+        <div>
+            <ArticleCard />
+        </div>
+    );
+};
+
+export default page;
